更新了ConvertTest.h作为0-49标准；实现了双击位置放置角色自动寻找（基础寻路）攻击建筑物的过程，Battletestlayer完全注释便于理解，添加了基本的战斗结束layer。

+++ Classes/BaseUnitData.h
@@ -41,8 +41,8 @@ struct BaseUnitData
     int level;                  //等级
 
     //战斗属性
-    int health;                 //生命值
-    int damage;                 //攻击力
+    float health;                 //生命值
+    float damage;                 //攻击力
     float attack_interval;     //攻击间隔（秒）
     float attack_distance;     //攻击距离（格）
